# vision_estate/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/newsletter/confirm": {
      "get": {
        "operationId": "NewsletterController_confirm",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 1024
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsletterResult"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Access denied",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Resource unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "State or concurrency conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected failure",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Database or integration unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "tags": [
          "Newsletter"
        ],
        "summary": "Consume a single-use newsletter confirmation token",
        "description": "Consume a single-use newsletter confirmation token",
        "security": []
      }
    },
...
  },
  "info": {
    "title": "Vision Estates API",
    "description": "Staff auth uses short-lived JWT and rotating HttpOnly refresh cookies with reuse detection. Legacy sessions expire naturally. Seller access uses a private header. Vendor integrations may be unavailable.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "ve_session"
      },
      "accessCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "ve_access"
      },
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "assessment": {
        "type": "apiKey",
        "in": "header",
        "name": "x-assessment-token"
      }
    },
    "schemas": {
...
      "NewsletterResult": {
        "type": "object",
        "properties": {
          "state": {
            "type": "string",
            "enum": [
              "CONFIRMED",
              "WITHDRAWN"
            ]
          }
        },
        "required": [
          "state"
        ]
      },
      "Problem": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ]
          },
          "instance": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance"
        ]
      },
...
    }
  }
}